+++ packages/deploy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/deploy
 
+## 0.7.0
+
+### Minor Changes
+
+- 0d53f9b: Add support for basic project-credential management (add, associate with jobs) via the CLI
+
 ## 0.6.0
 
 ### Minor Changes